+++ src/components/DogDetails-v2.js
