+++ src/views/Dogs/Dogs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useState, useEffect } from 'react';
 import { fetchDogs } from '../../services/dog';
-import DogList from '../../components/DogList';
+import DogList from '../../components/DogList-v2';
 
 export default function Dogs() {
   const [dogs, setDogs] = useState([]);
